14713: Migrate old crunch-dispatch-slurm config to new config
Update code to use new config.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <pamstutz@veritasgenetics.com>

+++ sdk/go/dispatch/dispatch.go
@@ -38,7 +38,7 @@ type Dispatcher struct {
 	Logger Logger
 
 	// Batch size for container queries
-	BatchSize int64
+	BatchSize int
 
 	// Queue polling frequency
 	PollPeriod time.Duration
